+++ libMx/mx_stdint.h
@@ -40,7 +40,7 @@ typedef short			int16_t;
 #  endif
 
 #  if !defined(HAVE_INT32_T)
-typedef long			int32_t;
+typedef int			int32_t;
 #     define HAVE_INT32_T
 #  endif
 
@@ -51,7 +51,7 @@ typedef long long		int64_t;
 
 typedef unsigned char		uint8_t;
 typedef unsigned short		uint16_t;
-typedef unsigned long		uint32_t;
+typedef unsigned int		uint32_t;
 typedef unsigned long long	uint64_t;
 
 /*---*/
